Fix clang warning.

+++ Cesium3DTilesSelection/src/upsampleGltfForRasterOverlays.cpp
@@ -150,7 +150,7 @@ Model upsampleGltfForRasterOverlays(
       // We're assuming here that nothing references primitives by index, so we
       // can remove them without any drama.
       if (!keep) {
-        mesh.primitives.erase(mesh.primitives.begin() + i);
+        mesh.primitives.erase(mesh.primitives.begin() + int64_t(i));
         --i;
       }
     }
